feat: Implement Main Menu (#6)
## Briefly your changes
Main menu with buttons to navigate to experiments, settings, help
(currently opens LabVerse repo in browser), and to quit the app. Panel
follows the user. No passthrough (want to test with headset). Added
tests for navigating to experiment selection scene and settings scene.
No tests for quitting application and opening documentation as they
could not be tested on my simulator, but they can be easily tested on
the actual device.

## Complete the below before requesting a review or merging
- [x] I have performed a self-review of my code.
- [x] If this is a feature branch, I have added thorough unit or
implementation tests.
- [x] I have ensured that all existing tests pass.

---------

Co-authored-by: Abdullah <abdullahijaz04@outlook.com>