feat: Experiment selection menu (#27)
## Briefly describe your changes
Selection Menu that lists the available experiments with descriptions
and preview images (currently placeholders that need to be replaced).
Has a back button to return to main menu and start button that changes
to ExperimentLab scene. Search bar that restricts options shown to ones
that match the search query.

## Complete the below before merging
- [x] I have performed a self-review of my code.
- [x] If this is a feature branch, I have added thorough unit or
implementation tests.
- [x] I have ensured that all existing tests pass.
- [x] My PR has been approved by at least one person that I've requested
to review.